aur-fetch: remove additional confirm_seen messages
This adds no new information over argv (if --confirm-seen is set,
AUR_SEEN is updated for all packages on the command-line) and --results
(distinction between cloned and fetched packages), and results in
copious output for large dependency chains. This holds in particular for
aur-sync, where all package names and their target versions are already
displayed on the console.